+++ CHANGELOG.md
@@ -64,6 +64,8 @@ Note that 3.1.0 will be the first API stable release and interfaces in this rele
 * The `seqan3::begin()`, `seqan3::end()`, `seqan3::cbegin()`, `seqan3::cend()`, `seqan3::size()`, `seqan3::empty()`
   functions have been deprecated:
   Use `std::ranges::{begin|end|cbegin|cend|size|empty}()` instead ([\#1663](https://github.com/seqan/seqan3/pull/1663)).
+* Added `seqan3::views::minimiser`, a view that computes the minimum in a window shifted over a range of comparable values.
+  ([\#1654](https://github.com/seqan/seqan3/pull/1654)).
 
 #### Search